fix l2 parameterization

+++ cosmobot_deep_learning/models/multi_image_cnn.py
@@ -175,7 +175,7 @@ def get_hyperparameter_parser():
     parser.add_argument("--dense-layer-units", type=int)
     parser.add_argument("--conv-dense-layer-1-units", type=int)
     parser.add_argument("--conv-dense-layer-2-units", type=int)
-    parser.add_argument("--l2-regularization", type=int),
+    parser.add_argument("--l2-regularization", type=float),
     parser.add_argument("--dropblock-size", type=int),
     parser.add_argument("--dropblock-keep-prob", type=float),
     parser.add_argument("--layer-activation", type=str),
